move encryption code to internal and other minor improvements
[#178795926](https://www.pivotaltracker.com/story/show/178795926)
