+++ pkg/broker/broker_test.go
@@ -22,6 +22,8 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/cloudfoundry-incubator/cloud-service-broker/internal/encryption"
+
 	"github.com/cloudfoundry-incubator/cloud-service-broker/db_service/models"
 	"github.com/cloudfoundry-incubator/cloud-service-broker/pkg/validation"
 	"github.com/cloudfoundry-incubator/cloud-service-broker/pkg/varcontext"
@@ -531,7 +533,7 @@ func TestServiceDefinition_ProvisionVariables(t *testing.T) {
 			}
 			defer viper.Reset()
 
-			models.SetEncryptor(models.NewNoopEncryptor())
+			models.SetEncryptor(encryption.NewNoopEncryptor())
 			details := domain.ProvisionDetails{RawParameters: json.RawMessage(tc.UserParams), RawContext: json.RawMessage(tc.RawContext)}
 			plan := ServicePlan{ServiceProperties: tc.ServiceProperties, ProvisionOverrides: tc.ProvisionOverrides}
 			vars, err := service.ProvisionVariables("instance-id-here", details, plan, tc.OriginatingIdentity)
@@ -1026,7 +1028,7 @@ func TestServiceDefinition_BindVariables(t *testing.T) {
 			viper.Set(service.BindDefaultOverrideProperty(), tc.DefaultOverride)
 			defer viper.Reset()
 
-			models.SetEncryptor(models.NewNoopEncryptor())
+			models.SetEncryptor(encryption.NewNoopEncryptor())
 			details := domain.BindDetails{RawParameters: json.RawMessage(tc.UserParams), RawContext: json.RawMessage(tc.RawContext)}
 			instance := models.ServiceInstanceDetails{}
 			instance.SetOtherDetails(tc.InstanceVars)
